+++ crates/infra/orpc-rpc/Cargo.toml
@@ -6,3 +6,14 @@ license.workspace = true
 description = "ORPC RPC migration boundary."
 
 [dependencies]
+curvine-io = { workspace = true }
+curvine-sys = { workspace = true }
+orpc-error = { workspace = true }
+orpc-net = { workspace = true }
+orpc-runtime = { workspace = true }
+bytes = { workspace = true }
+tokio = { workspace = true }
+tokio-util = { workspace = true }
+prost = { workspace = true }
+num_enum = { workspace = true }
+log = { workspace = true }
