+++ crates/infra/orpc-rpc/src/handler/stream_handler.rs
@@ -13,11 +13,11 @@
 // limitations under the License.
 
 use crate::handler::{Frame, MessageHandler};
-use crate::io::IOResult;
 use crate::message::{Builder, Message};
-use crate::runtime::{RpcRuntime, Runtime};
-use crate::server::ServerConf;
+use crate::ServerConf;
+use curvine_io::IOResult;
 use log::debug;
+use orpc_runtime::runtime::{RpcRuntime, Runtime};
 use std::sync::Arc;
 use std::time::Duration;
 use tokio::time::timeout;
