+++ crates/infra/orpc-rpc/src/handler/write_frame.rs
@@ -16,8 +16,8 @@ use tokio::io::{AsyncWriteExt, WriteHalf};
 use tokio::net::TcpStream;
 
 use crate::handler::FrameBuf;
-use crate::io::IOResult;
 use crate::message::Message;
+use curvine_io::IOResult;
 
 pub struct WriteFrame {
     io: WriteHalf<TcpStream>,
