+++ crates/infra/orpc-rpc/src/lib.rs
@@ -12,4 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// RPC codec/client/server primitives will migrate here from the compatibility facade.
+#[macro_use]
+mod macros;
+
+pub mod handler;
+pub mod message;
+mod server_conf;
+
+pub use server_conf::ServerConf;
