+++ crates/infra/orpc-rpc/src/message/message_builder.rs
@@ -12,13 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::common::Utils;
 use crate::message::{
     Builder, Message, Protocol, RequestStatus, ResponseStatus, Status, EMPTY_REQ_ID, INIT_SEQ_ID,
 };
-use crate::sys::DataSlice;
 use bytes::BytesMut;
+use curvine_io::DataSlice;
 use log::error;
+use orpc_runtime::common::Utils;
 use prost::Message as PMessage;
 use std::fmt::Debug;
