+++ orpc/Cargo.toml
@@ -12,6 +12,7 @@ curvine-io = { workspace = true }
 curvine-sys = { workspace = true }
 orpc-error = { workspace = true }
 orpc-net = { workspace = true }
+orpc-rpc = { workspace = true }
 orpc-runtime = { workspace = true }
 tokio = { workspace = true }
 futures = { workspace = true }
